pyproject.toml
[project]
name = "hostelops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
